backend/requirements.txt

flask==2.3.3
flask-cors==4.0.0
neo4j==5.13.0
python-dotenv==1.0.0
